requirements.txt
# File: backend/requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
numpy
pillow
torch
torchvision